pyproject.toml
[project]
name = "tgpr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "albumentations>=2.0.8",
    "einops>=0.8.1",
    "faiss-cpu>=1.13.0",
    "fastai>=2.8.5",
    "geoopt>=0.5.1",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "lightgbm>=4.6.0",
    "lightning>=2.5.6",
    "matplotlib>=3.10.7",
    "notebook>=7.4.7",
    "numpy>=2.3.5",
    "open-clip-torch>=3.2.0",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "pillow>=12.0.0",
    "pre-commit>=4.4.0",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "pytorch-metric-learning>=2.9.0",
    "rich>=14.2.0",
    "rootutils>=1.0.7",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchmetrics>=1.8.2",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "wandb>=0.23.0",
    "xgboost>=3.1.2",
]
